lesson1

+++ src/lesson03/Task4.java
@@ -1,8 +1,8 @@
 package lesson03;
 
 public class Task4 {
-    public int sum(int mon, int tus, int wen, int thu, int fr, int sat, int sun) {
-        int result = mon + tus + wen + thu + fr + sat + sun;
+    public double sum(double mon, double tus, double wen, double thu, double fr, double sat, double sun) {
+        double result = mon + tus + wen + thu + fr + sat + sun;
 
         return result;
     }
